requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai>=1.35.7
tiktoken>=0.7.0
chromadb>=0.5.3
pyyaml>=6.0.1
pydantic>=2.7.4
httpx>=0.27.0
python-dotenv>=1.0.1
